Add files via upload
Abstract—In the current world, transportation is considered
as one of the basic requirements in humans day
to day life. Air travel has revolutionized transportation
industry and changed the mindset of human that
traveling to any part of the world isnt that difficult.
This massive improvement also brings in many practical
difficulties to this industry, especially factors like flight
delays, flight cancellations, and operations of flights by
the airport. These problems are analyzed, identified and
being fixed. But with the generation of a huge amount of
data, it has become a struggle to store and analyze the
data such data. In this project, a massive volume of flight
operation data in USA has been gathered and processed
in big data environments using HDFS architecture and
databases like MySQL, Hive, HBase. Sqoop has been
to successfully transfer the data from one database to
another. MapReduce programs and Hive queries were
used to analyze the flight data to address some of the
problem areas in the aviation industry which could bring
in more efficiency in their operation.